Registeration and login are done
Registeration function with handling everything
