--- Users.txt
@@ -1,2 +1,0 @@
-user 1,ahmed,12345,JetStream Sam
-user 2,ziad,1234,Zoldikxx
